# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小知，你的校园助手",
        "title": "KnowEase API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/{type}/{postid}/{userid}/save": {
            "post": {
                "description": "用户给帖子收藏，收藏数加一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-收藏"
                ],
                "summary": "用户收藏帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "帖子类型",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户收藏成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "收藏记录上传失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/getcounts": {
            "get": {
                "description": "获取帖子的点赞数，收藏数，浏览量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子"
                ],
                "summary": "获取帖子相关数值",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/getstatus": {
            "get": {
                "description": "获取帖子点赞收藏状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区"
                ],
                "summary": "获取帖子点赞收藏状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子状态信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/{commentid}/getcounts": {
            "get": {
                "description": "获取评论的点赞数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "获取评论点赞数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}/{commentid}/getstatus": {
            "get": {
                "description": "获取评论点赞状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区"
                ],
                "summary": "获取评论点赞状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评论状态信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/{commentid}/{replyid}/getcounts": {
            "get": {
                "description": "获取回复的点赞数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子-评论"
                ],
                "summary": "获取回复点赞数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "回复ID",
                        "name": "replyid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "相关数值",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}